src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/reference/ArrayCharacterAttributeParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.reference;

import com.github.brunomndantas.jscrapper.support.parser.array.attribute.ArrayAttributeParser;
import org.openqa.selenium.WebElement;

public class ArrayCharacterAttributeParser extends ArrayAttributeParser {

    protected ArrayCharacterAttributeParser(String attribute) {
        super(attribute, null);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return element.getAttribute(super.attribute).charAt(0);
    }

}
