src/test/java/com/github/brunomndantas/jscrapper/support/parser/array/attribute/reference/ArrayCharacterAttributeParserTest.java
package com.github.brunomndantas.jscrapper.support.parser.array.attribute.reference;

import com.github.brunomndantas.jscrapper.DummyElement;
import org.junit.Test;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertSame;

public class ArrayCharacterAttributeParserTest{

    @Test
    public void constructorTest() {
        String attribute = "";
        ArrayCharacterAttributeParser parser = new ArrayCharacterAttributeParser(attribute);
        assertEquals(null, parser.getDefaultValue());
        assertSame(attribute, parser.getAttribute());
    }

    @Test
    public void parseElement() throws Exception {
        ArrayCharacterAttributeParser parser = new ArrayCharacterAttributeParser("");

        assertEquals('\u0001', parser.parseElement(new DummyElement(){
            @Override
            public String getAttribute(String a) {
                return ((char)1) + "";
            }
        }));
    }

}
